Improve naming of `QueryValidation` hooks.
